Add a catch-all argument parsing strategy (#29)
* Add a catch-all argument parsing strategy

We don't currently have a way for `@Argument` arrays to capture
command-line inputs that look like options. This capability is
important for tools like SwiftPM that need to forward input to
another command.

This introduces an `ArgumentArrayParsingStrategy` enum with a
`remaining` case that  matches the current behavior, as well as an
`unconditionalRemaining` case that captures all remaining input. The
two array-based `@Argument` initializers gain defaulted parameters for
the parsing strategy.

* Address feedback from @danieleggert

- Added a bit of a warning for the unconditional parsing strategy
- Switched to a better way of getting the original input
- Changed to pulling the "earliest" element when combining positional and
  unused values, rather than combining and sorting. This felt more
  efficient than reconstructing a SplitArguments instance or combining
  and sorting.

* Add parameter docs for argument parsing strategy

* Expand documentation and tests to cover `--` terminator

* Fix case name in API documentation