Work around runtime issue introduced in 0.4.3 (#342)
* Add dummy property as workaround for #338

The addition of _hiddenFromHelp to OptionGroup in 0.4.3 triggered
a Swift IRGen issue. Adding this dummy variable works around the
issue in at least some cases.

* Don't include _hiddenFromHelp props in other property wrappers

This resolves another crashing issue described in rdar://80796582,
where defining an OptionGroup that isn't used results in a runtime
segfault when compiled in release mode.