fix: 🐛 Declare `GenerateManualPlugin` as plugin product (#456)
`ArgumentParser` 1.1.3 includes a SwiftPM plugin for generating man
pages. However, this plugin is not visible to packages that declare
`ArgumentParser` as a package dependency.

Defining a product of type `plugin` in the package manifest will solve
that. Running `swift package plugin --list` for a package, that
declares `ArgumentParser` as a package dependency, will then list the
`GenerateManualPlugin`

```
‘experimental-generate-manual’ (plugin ‘GenerateManualPlugin’ in package ‘swift-argument-parser’)
```

Closes: #455