Don't remove nested option group titles (#592)
When option groups are nested, any titles of the nested groups are
removed, even when the containing group doesn't have a title. This
change preserves the nested groups' titles when grouped in an
untitled option group. Nesting within a titled option group continues
to override any nested option groups.

Fixes #585.