Abort on ambiguous subcommand names
Currently it's possible to have two commands with the same _commandName
by mixing a combination of the automatic command name generation based on
the type name, and explicit use of `CommandConfiguration`'s commandName
parameter. For example, if we edit the math examples `Add` command to
have a commandName of "multiply" this is allowed currently:

```
./.build/debug/math multiply 2 3
5
```

The behavior today is whatever subcommand is registered in the tree first
is what will resolve for that command. So, `Multiply` in this example is
permanently shadowed.

This change just makes sure there's no occurence of this happening in any
level of the tree of potential subcommands, and if so we'll abort early.