Add ability to retrieve list of possible options for Option property
An EnumerableOption protocol has been created to allow users to specify
help/discussion information on a custom Option type itself. This
information can be accessed further down the pipeline, which permits the
retrieval of all the option types and their descriptions for dump/help
generation.

Note that this has yet to be formatted and is currently WIP.