Address PR comments and fix option label formatting
If an option case's label was too long for the description
to fit on the same line, it would omit the label entirely.

Append the label to the rendered help text with a newline if
the label is too long for the description to fit on the same
line.

Added test cases to address scenarios where the option cases
either have:
* a label that is too long for the description to fit on the
  same line
* a description that needs to be wrapped
* a label that is too long and a description that needs to be
  wrapped
* all of the above cases when an Option also has a discussion,
  which further indents the list of possible option values in
  the rendered help text.