Escape backslashes for fish single-quoted strings.
If someone already escapes backslashes for such strings, this will cause an issue, but no one should be required to go to the trouble to manually escape backslashes in their strings, especially since the requirement isn't documented or normal.

If someone doesn't escape backslashes, then that can break then old version of the script.

Signed-off-by: Ross Goldberg <484615+rgoldberg@users.noreply.github.com>