Updates output to present the string representation of command arguments to a wrapped text block (#791)
Arguments switched from a single long line within a preformatted text
block to multilines and wrapped to a line length. This makes the output
significantly easier to read when there's a large number of options
(for example, the `swift build` command). The wrapping algorithm is
a simple greedy algorithm that extends up to, but not beyond, a
maximum text width.

The line length defined by the style of output (markdown - 80, docc - 60)
- Snapshot tests for output updated to reflect changes