Introduce tool info version 1
Fork the tool info version 0 to a new stable version 1. Introduce
a new '--help-dump-tool-info-v1' flag that functions similarly to
'--experimental-dump-help'. Create a new ToolInfoV1 to keep the
new types for the new schema. Update the tool info serialization
version to 1. Add a JSON schema for the version 1 tool info
format.